fixed linter and tests so they are now passing and working

+++ config/application.rb
@@ -28,9 +28,11 @@ module DeviseVue
     # config.eager_load_paths << Rails.root.join("extras")
     config.before_configuration do
       env_file = File.join(Rails.root, 'config', '.env.yml')
-      YAML.load(File.open(env_file)).each do |key, value|
-        ENV[key.to_s] = value
-      end if File.exists?(env_file)
+      if File.exist?(env_file)
+        YAML.load(File.open(env_file)).each do |key, value|
+          ENV[key.to_s] = value
+        end
+      end
     end
     # Only loads a smaller set of middleware suitable for API only apps.
     # Middleware like session, flash, cookies can be added back manually.
